+++ modules/sdk/gradle-plugins-defaults/CHANGELOG.markdown
@@ -224,11 +224,162 @@ directory instead of the root project directory. Doing this lets submodules like
 ## 1.2.21 - 2016-09-13
 
 ### Changed
-- [LRDOCS-2981]: Prepend `"Module "` to `appJavadoc` module headings.
+- [LRDOCS-2981]: Prepend *Module* string to `appJavadoc` module headings.
+
+## 1.2.22 - 2016-09-14
+
+### Changed
+- [LPS-68131]: Update the [Liferay Gradle Plugins] dependency to version 2.0.34.
+
+## 1.2.23 - 2016-09-16
+
+### Changed
+- [LPS-68131]: Update the [Liferay Gradle Plugins] dependency to version 2.0.35.
+
+## 1.2.24 - 2016-09-20
+
+### Changed
+- [LPS-67352]: Update the [Liferay Gradle Plugins] dependency to version 2.0.36.
+- [LPS-68230]: Configure [Liferay Gradle Plugins Node] to use version 6.6.0 of
+Node.js.
+
+### Removed
+- [LPS-68230]: To reduce the number of plugins applied to a project and improve
+performance, plugins in `com.liferay.gradle.plugins.defaults.internal` are no
+longer applied via `apply plugin`.
+
+## 1.2.25 - 2016-09-20
+
+### Changed
+- [LPS-66906]: Update the [Liferay Gradle Plugins] dependency to version 2.0.37.
+
+## 1.2.26 - 2016-09-21
+
+### Changed
+- [LPS-66853]: Update the [Liferay Gradle Plugins] dependency to version 2.0.38.
+
+## 1.2.27 - 2016-09-22
+
+### Changed
+- [LPS-68305]: Set the `buildService.buildNumberIncrement` property to `false`
+by default.
+
+## 1.2.28 - 2016-09-22
+
+### Changed
+- [LPS-68297]: Update the [Liferay Gradle Plugins] dependency to version 2.0.39.
+
+## 1.2.29 - 2016-09-22
+
+### Added
+- [LPS-66906]: Override the [`sass-binary-path`](https://github.com/sass/node-sass#binary-configuration-parameters)
+argument in the `npmInstall` task with the value of the project property
+`nodejs.npm.ci.sass.binary.site` when using Jenkins.
+
+### Changed
+- [LPS-66906]: Update the [Liferay Gradle Plugins] dependency to version 2.0.40.
+
+## 1.2.30 - 2016-09-23
+
+### Changed
+- [LPS-66853]: Update the [Liferay Gradle Plugins] dependency to version 2.0.41.
+
+## 1.2.31 - 2016-09-23
+
+### Added
+- [LPS-68306]: Set the system property `portal.pre.build` to `true` to only
+include the projects containing a `.lfrbuild-portal-pre` marker file.
+
+## 1.2.32 - 2016-09-26
+
+### Changed
+- [LPS-66853]: Update the [Liferay Gradle Plugins] dependency to version 2.0.42.
+
+## 1.2.33 - 2016-09-27
+
+### Changed
+- [LPS-66853]: Update the [Liferay Gradle Plugins] dependency to version 2.0.44.
+
+## 1.2.34 - 2016-09-27
+
+### Changed
+- [LPS-67863]: Change dependency version override declarations in
+`.version-override-${project.name}.properties` to follow a new format:
+
+		[artifact group]/[artifact name]=[new version]
+
+### Fixed
+- [LPS-67863]: Fix commit deletion process of version override files.
+
+## 1.2.35 - 2016-09-28
+
+### Changed
+- [LPS-67352]: Update the [Liferay Gradle Plugins] dependency to version 2.0.45.
+
+## 1.2.36 - 2016-09-29
+
+### Changed
+- [LPS-58672]: Update the [Liferay Gradle Plugins] dependency to version 2.0.46.
+
+## 1.2.37 - 2016-09-30
+
+### Changed
+- [LPS-67352]: Update the [Liferay Gradle Plugins] dependency to version 2.0.47.
+
+## 1.2.38 - 2016-10-01
+
+### Added
+- [LPS-68448]: Fail the build of an OSGi project if the version in the
+`npm-shrinkwrap.json` file does not match the project version.
+- [LPS-68448]: The task `updateVersion` of OSGi and theme projects updates the
+version in the `npm-shrinkwrap.json` file, if present.
+
+## 1.2.39 - 2016-10-03
+
+### Added
+- [LPS-68402]: Set the [`org.apache.maven.offline`](https://github.com/shrinkwrap/resolver#system-properties)
+system property to `true` for the `testIntegration` task.
+
+### Changed
+- [LPS-68485]: Update the [Liferay Gradle Plugins] dependency to version 2.0.48.
+
+## 1.2.40 - 2016-10-04
+
+### Added
+- [LPS-68506]: Exclude unpublished projects from the API documentation generated
+by the `appJavadoc` task.
+
+### Changed
+- [LPS-68504]: Update the [Liferay Gradle Plugins] dependency to version 2.0.49.
+- [LPS-68506]: Update the [Liferay Gradle Plugins App Javadoc Builder]
+dependency to version 1.1.0.
+
+## 1.2.41 - 2016-10-05
+
+### Added
+- [LPS-68540]: Fail the `uploadArchives` task execution if the project directory
+contains the marker file `.lfrbuild-missing-resources-importer`.
+
+## 1.2.42 - 2016-10-05
+
+### Added
+- [LPS-66396]: By setting the following system properties to `true`, it is now
+possible to exclude specific project types from the build:
+	- `build.exclude.ant.plugin` to exclude all the projects that contain a `build.xml` file from the build
+	- `build.exclude.module` to exclude all the projects that contain a `bnd.bnd` file from the build
+	- `build.exclude.theme` to exclude all the project that contain a `gulpfile.js` file from the build
+
+### Removed
+- [LPS-66396]: The `modules.only.build` system property is no longer available.
 
 [Liferay Gradle Plugins]: https://github.com/liferay/liferay-portal/tree/master/modules/sdk/gradle-plugins
+[Liferay Gradle Plugins App Javadoc Builder]: https://github.com/liferay/liferay-portal/tree/master/modules/sdk/gradle-plugins-app-javadoc-builder
+[Liferay Gradle Plugins Node]: https://github.com/liferay/liferay-portal/tree/master/modules/sdk/gradle-plugins-node
+[LPS-58672]: https://issues.liferay.com/browse/LPS-58672
 [LPS-61099]: https://issues.liferay.com/browse/LPS-61099
+[LPS-66396]: https://issues.liferay.com/browse/LPS-66396
 [LPS-66853]: https://issues.liferay.com/browse/LPS-66853
+[LPS-66906]: https://issues.liferay.com/browse/LPS-66906
 [LPS-67023]: https://issues.liferay.com/browse/LPS-67023
 [LPS-67352]: https://issues.liferay.com/browse/LPS-67352
 [LPS-67658]: https://issues.liferay.com/browse/LPS-67658
@@ -240,5 +391,16 @@ directory instead of the root project directory. Doing this lets submodules like
 [LPS-67996]: https://issues.liferay.com/browse/LPS-67996
 [LPS-68009]: https://issues.liferay.com/browse/LPS-68009
 [LPS-68014]: https://issues.liferay.com/browse/LPS-68014
+[LPS-68131]: https://issues.liferay.com/browse/LPS-68131
+[LPS-68230]: https://issues.liferay.com/browse/LPS-68230
+[LPS-68297]: https://issues.liferay.com/browse/LPS-68297
+[LPS-68305]: https://issues.liferay.com/browse/LPS-68305
+[LPS-68306]: https://issues.liferay.com/browse/LPS-68306
+[LPS-68402]: https://issues.liferay.com/browse/LPS-68402
+[LPS-68448]: https://issues.liferay.com/browse/LPS-68448
+[LPS-68485]: https://issues.liferay.com/browse/LPS-68485
+[LPS-68504]: https://issues.liferay.com/browse/LPS-68504
+[LPS-68506]: https://issues.liferay.com/browse/LPS-68506
+[LPS-68540]: https://issues.liferay.com/browse/LPS-68540
 [LRDOCS-2841]: https://issues.liferay.com/browse/LRDOCS-2841
 [LRDOCS-2981]: https://issues.liferay.com/browse/LRDOCS-2981